+++ odoo/local-src/sf_rma/models/rma_config_settings.py
@@ -1,0 +1,15 @@
+# -*- coding: utf-8 -*-
+# Copyright 2017 Camptocamp SA
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+
+from odoo import fields, models
+
+
+class RMASettings(models.TransientModel):
+
+    _name = 'rma.config.settings'
+    _inherit = 'res.config.settings'
+
+    company_id = fields.Many2one(
+        'res.company', string='Company', required=True,
+        default=lambda self: self.env.user.company_id)