+++ odoo/local-src/sf_rma_sale_order/models/__init__.py
@@ -1,0 +1,2 @@
+from . import res_company
+from . import rma_config_settings